+++ CHANGELOG.md
@@ -5,7 +5,8 @@ This file is used to list changes made in each version of the windows cookbook.
 Unreleased
 --------------------
 - fix `windows_package` `Chef.set_resource_priority_array` warning
-- update windows_task to support tasks in folders
+- update `windows_task` to support tasks in folders
+- fix `windows_task` delete action
 
 v1.36.6 (2014-12-18)
 --------------------